+++ lib/syntax_search/display_code_with_line_numbers.rb
@@ -24,33 +24,48 @@ module SyntaxErrorSearch
     TERMINAL_END = "\e[0m"
 
     def initialize(lines: , highlight_lines: [], terminal: false)
-      @lines = lines.sort
+      @lines = Array(lines).sort
       @terminal = terminal
-      @highlight_line_hash = highlight_lines.each_with_object({}) {|line, h| h[line] = true  }
+      @highlight_line_hash = Array(highlight_lines).each_with_object({}) {|line, h| h[line] = true  }
       @digit_count = @lines.last&.line_number.to_s.length
     end
 
     def call
       @lines.map do |line|
-        string = String.new("")
-        if @highlight_line_hash[line]
-          string << "❯ "
-        else
-          string << "  "
-        end
-
-        number = line.line_number.to_s.rjust(@digit_count)
-        string << number.to_s
-        if line.empty?
-          string << line.original_line
-        else
-          string << "  "
-          string << TERMINAL_HIGHLIGHT if @terminal && @highlight_line_hash[line] # Bold, italics
-          string << line.original_line
-          string << TERMINAL_END if @terminal
-        end
-        string
+        format_line(line)
       end.join
+    end
+
+    private def format_line(code_line)
+      # Handle trailing slash lines
+      code_line.original.lines.map.with_index do |contents, i|
+        format(
+          empty: code_line.empty?,
+          number: (code_line.number + i).to_s,
+          contents: contents,
+          highlight: @highlight_line_hash[code_line]
+        )
+      end.join
+    end
+
+    private def format(contents: , number: , highlight: false, empty:)
+      string = String.new("")
+      if highlight
+        string << "❯ "
+      else
+        string << "  "
+      end
+
+      string << number.rjust(@digit_count).to_s
+      if empty
+        string << contents
+      else
+        string << "  "
+        string << TERMINAL_HIGHLIGHT if @terminal && highlight
+        string << contents
+        string << TERMINAL_END if @terminal
+      end
+      string
     end
   end
 end